refactor(credit_card): credit card 앱 분리
