+++ apps/credit_card/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class CreditCardConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'apps.credit_card'
