+++ apps/order/models.py
@@ -1,7 +1,8 @@
 import typing
 from django.db import models
 from pangpangeats.settings import AUTH_USER_MODEL
-from apps.user.models import CreditCard, User
+from apps.user.models import User
+from apps.credit_card.models import CreditCard
 from apps.restaurant.models import MenuInformation
