+++ pangpangeats/settings.py
@@ -41,6 +41,7 @@ INSTALLED_APPS = [
     'drf_spectacular',
     # my apps
     'apps.user',
+    'apps.credit_card',
     'apps.restaurant',
     'apps.order',
 ]
